+++ packages/app/src/vite-env.d.ts
@@ -8,6 +8,7 @@ interface ImportMetaEnv {
   readonly VITE_CRX: string;
   readonly VITE_ADDR_OWNER: string;
   readonly VITE_CRX_NAME: string;
+  readonly VITE_AUTO_LOCK_IN_MINUTES: number;
 }
 
 interface ImportMeta {